refactor(race): 5단계 - 자동차 경주(리팩토링)
- domain, view, controller 로 패키지 구조 변경
- 불필요한 CarRandomMoveRuleValue 삭제
- 전략 패턴에 맞춰 테스트 코드, 네이밍 변경
