--- src/main/java/race/CarMoveRule.java
@@ -1,5 +1,0 @@
-package race;
-
-public interface CarMoveRule {
-    boolean check(CarMoveRuleValue value);
-}
