--- src/main/java/race/CarMoveRuleValue.java
@@ -1,5 +1,0 @@
-package race;
-
-public interface CarMoveRuleValue {
-    boolean isGreaterThanOrEqualTo(int value);
-}
