+++ src/main/java/race/domain/CarMoveRule.java
@@ -1,0 +1,5 @@
+package race.domain;
+
+public interface CarMoveRule {
+    boolean movable();
+}
