--- src/test/java/race/MockCarMoveRule.java
@@ -1,8 +1,0 @@
-package race;
-
-public class MockCarMoveRule implements CarMoveRule {
-    @Override
-    public boolean check(CarMoveRuleValue value) {
-        return value instanceof MockForwardCarMoveRuleValue;
-    }
-}
