--- src/test/java/race/MockForwardCarMoveRuleValue.java
@@ -1,8 +1,0 @@
-package race;
-
-public class MockForwardCarMoveRuleValue implements CarMoveRuleValue {
-    @Override
-    public boolean isGreaterThanOrEqualTo(int value) {
-        return true;
-    }
-}
