--- src/test/java/race/MockHoldCarMoveRuleValue.java
@@ -1,8 +1,0 @@
-package race;
-
-public class MockHoldCarMoveRuleValue implements CarMoveRuleValue {
-    @Override
-    public boolean isGreaterThanOrEqualTo(int value) {
-        return true;
-    }
-}
